[Utility]: Added an IOP file reader
Added a simple way to read IOP files to assist with testing.
Feel free to use this in applications that consume this library to easily read an IOP file, but it may not be ideal for an embedded platform with very limted RAM, as it will store the entire IOP file content in RAM.
If you need to read an object pool in segments due to RAM restrictions, you should instead use CAN data chunk callbacks in concert with your own custom driver for getting pool data.
If there is demand for it, I may add an interface for that in the future.