[TP]: Fixed several issues with transmitting a BAM session
I found that transmitting a BAM session was missing several components while working on the DM1 and DM2 messages.
In particular, sending the actual BAM itself was missing, the frame timing for BAM was missing, and I was using some bad assumptions in the code from CM.
This commit allows BAM sessions to actually send and complete.