[Core]: Refactor hardware abstraction layer to better modularize around drivers
Broke-up the socket-can layer to be a more generic abstraction layer.
Now, if you want to implement your own driver for some other CAN interface, you just need to inherit from `CANHardwarePlugin` and write the 5 functions needed.
Also added a way to not build the examples. I need to still add some CMake to not compile the examples if SocketCAN is not present and add some CMake to allow selecting other CAN drivers.