[Core]: Added a way to register for a PGN callback for any non-partner control function
This is is essentially a way to register for PGN callbacks from anybody.
This should really only be used when the other registration method doesn't work for your use case, but for some server applications (VT and File Server come to mind) it seems like it would be beneficial to have a way to get all messages with a specific PGN without registering specific partners.