[#102]: Change how partner and ICF list maintenance is done when members are deleted
When a partner or ICF is deleted, now we will null out its pointer in the list rather than erasing it, which could have caused an exception if the deleted object was a created via smart pointer.
Likewise, when a new object is created, we will search first for empty spots in the list before allocating new ones to keep the list as small as possible.
Added a trivial unit test to detect exceptions on creation and deletion of ICFs and Partners.