[Core]: Fix the sonar blocker items
Fixed issue in TP similar to the one I fixed in #141
Fixed a check where the scanner was worried about class slicing (but no slicing was probably happening?)
Add a condition to the can thread wait. We really just want it to always wake up, as there are tons of things that trigger that thread but guidelines say we should have some condition.