[VT]: Fix auto-scale crash when using a pool version label
Changed when auto-scaling happens to avoid trying to upload from an empty scaling buffer.
Made some of the scaling functions static where possible.
Changed one conditional to an assert to make it clearer to a user when they call set_object_pool_scaling with out a valid pool.
Capitalized the 't' in "VT" in the VT client's test implementation.