[Core]: Mitigate an order of operations issue in control function destructors when control functions are created as global static objects
Sometimes the order of static object destruction can be a bit unpredictable, so sometimes I was seeing that
the lists that contain the internal and partner control functions would be cleared before we could access them in the destructors
of those objects. So I added a trivial check to ensure those lists contains > 0 objects before trying to do any further operations with it
inside those destructors.