[Core]: Added a way to disable the hardware plugin receive threads
I was having a lot of trouble while testing task controller message encoding because the Rx thread for the virtual can plugin could come along and read messages out of the queues before I could manually read them to verify their content.
So, these changes allow us to temporarily stop the Rx threads so that we can synchronously receive messages while running tests to ensure we never randomly lose messages (which we were before).
I also added a timeout value to the virtual can read function to 100% avoid an infinite wait. Generally this will not affect anything as the Rx thread will just re-try the read, but it has the upside of preventing a hard lock up while running tests if a message fails to send or something.