[#245]: Moved NMEA2k fast packet protocol instance to the network manager
Moved the fast packet protocol instance into the network manager to prevent it from getting optimized out when linking as a static library.
Added a getter to the network manager to get the protocol instance.
Fixed one session variable not being initialized at construction (though this probably was not causing issues).
Updated the FP example to use the new protocol instance.