[ISB]: Changed ISB to be updated from the app instead of by the stack
This generally avoids a race condition between construction of an
ISB by the application and an update by the network manager.
This could also be avoided by careful timing of how functions are
called when initializing in the app, but since it's an implement
application layer message, it probably makes more sense to have the user
update it from their application.
Fixes a random crash in the unit tests.