[TC]: Add options for falling back to other language data sources when TC doesn't respond
Allow the TC client to fall back to other language sources as needed.
Fixed a condition where the TC client would become stuck while waiting for a language response if a response never arrived.
Added a getter to the language command interface to allow retrieving the assigned partner CF.
Prevent a crash in the language command interface which could occur if you send the command with some fields left unconfigured.